guessing game
these are simple programs that guesses the number between 1 to n. the
guessing game has a human input where a user guesses a number from the
computer or other human and a computer guesses the users number choice
by using a random generator. the other file the computer guesses your
number using a binary search algorithm.